upgrade b4.5

+++ MeadowGameofLife/MeadowApp.cs
@@ -55,7 +55,7 @@ namespace MeadowGameofLife
                 chipSelectPin: null,
                 dcPin: Device.Pins.D01,
                 resetPin: Device.Pins.D00,
-                width: (uint)displayWidth, height: (uint)displayHeight);
+                width: displayWidth, height: displayHeight);
 
             onboardLed = new RgbPwmLed(device: Device,
                 redPwmPin: Device.Pins.OnboardLedRed,
